Fix crash in puzzles with alternate clue lists.
Fixes #11.

When a puzzle uses something other than "Across" and "Down", e.g. a
marching bands puzzle with "Bands" and "Rows", we would update the
pane captions but not the corresponding menu items in View -> Panes.
This causes a crash when MyAuiMgr attempts to look up the pane for the
stale menu item "Across" and fails to match it by caption with the
pane for "Bands".

So, whenever we update the clue lists, update the menu as well.