Suppress grid clicks while "Paused" is rendered.
Fixes #13.

If focus is elsewhere and the grid is showing the "Paused" screen,
clicks should not be processed - the user cannot actually see where in
the grid they're clicking, so this can result in a disorienting change
in focus. The puzzle should be resumed exactly where the user left
off.

Note that it is not sufficient to check m_isPaused in OnLeftDown,
because the window activation event is processed before the mouse
click event. So we add a separate flag which reflects whether the
"Paused" screen is currently being rendered and only update this in
OnPaint when the user's view has actually changed.