Skip empty clue lists when loading JPZ files.
Fixes #22.

This fixes loading of Outside the Box Year 2 Variety Puzzle #2.
This JPZ file has one clue list, but (presumably) for compatibility
with apps that expect two lists, it specifies a second, empty list
of clues. Just ignore such lists and continue parsing. If all lists
are empty, detect this and throw a different error.