Fix grid control preference and OS X preference saving.
Fixes #3.

When using MOVE_TO_NEXT_BLANK, MOVE_AFTER_LETTER must also be set, or
else the setting is ignored and the preference reset to unchecked when
the app is restarted.

Reverts commit 3670eff, which moved defines from hpp to cpp, because
PreferencesPanel.hpp depends on the value of
XWORD_PREFERENCES_LIVE_PREVIEW, which was undefined after this change.
We must define this variable somewhere accessible to both
PreferencesDialog.cpp and PreferencesPanel.hpp. This fixes live saving
of settings changes on OS X.