Configure Travis and AppVeyor builds.
This is a start towards resolving Issue #40. These configs build the
debug and release builds of XWord on Mac and Windows. If a new release
is created on GitHub, a release build will be zipped up and posted as
part of the release.

The --enable-unicode reference in README.md is removed, as this is the
default as of wxWidgets 2.9.