Use cross-platform API to show Lua errors.
For debug builds, if any Lua errors occur, we try to show the log file
when the app is closed. However, this is done by trying to execute the
file directly, which doesn't work on Mac/Linux. Instead, use a
cross-platform wxWidgets API to open the file in the default handler.