Don't pass wxString::c_str() to vararg functions.
See the following note from a comment in string.h:

===
One of the places where the conversion does *NOT* happen correctly is
when c_str() is passed to a vararg function such as printf() so you
must *NOT* use c_str() with them. Either use wxPrintf() (all wx
functions do handle c_str() correctly, even if they appear to be
vararg (but they're not, really)) or add an explicit AsChar() or, if
compatibility with previous wxWidgets versions is important, add a
cast to "const char *".
===

Fixes #21