Work around malformatted [1] diagramless puzzles.
If a puzzle has:

-The "diagramless" bit set to true
-No timer information or the timer set to 0
-A user grid with all the black squares perfectly entered by the
"user" but no letters entered

Then it is highly likely to be malformatted, so replace all the
user-entered black squares with blanks as though the user's section
were filled with ':' instead of '.'. This prevents the grid from being
immediately visible upon opening such a puzzle, which spoils the
diagramless aspect entirely.

[1] https://twitter.com/thatpuzzleguy/status/709389211493752832