Generalize diagramless workaround and add escape latch.
Move the workaround for malformatted diagramless puzzles - those
with black squares entered instead of blank squares - to the
general puzzle loading path so it benefits all formats instead of
just .puz.

Also introduce a Tools -> Diagramless -> Convert to Normal menu
item which converts a diagramless puzzle into a normal one by
revealing all black squares. In addition to being a reasonable
feature on its own - perhaps a user just wants to solve the
puzzle normally, or can't figure out the diagram and wants to
reveal it - this provides an escape hatch where the workaround
was somehow erroneously applied.