Clean up Windows hack in paths.cpp.
Instead of working around the Windows-specific behavior of using the
parent directory of the .exe instead of the directory containing the
.exe itself when the .exe directory contains "debug", we can turn off
this behavior by calling DontIgnoreAppSubDir().

An alternative would be to force on portable mode for debug builds,
but this has other side effects which may not be desirable, e.g.
using a different config/settings from the release build.

Fixes #41