Make py2lua more robust in a git environment.
Git doesn't save mtimes, so relying on mtime alone to determine
whether the input wxFB.py file has changed can lead to spurious
wxFB.lua regeneration. To mitigate this, store the md5sum of the input
file in the generated lua code as well, and check this in addition
when the mtimes indicate a possible change to the input.

If the md5sums are still identical, we touch the python file to put
things back in a reasonable state (so we don't have to calculate a
hash of the input file on every startup).

MD5 implementation via https://github.com/kikito/md5.lua