Fix two-finger scrolling in modals on Mac.
This appears to be a bug in wxWidgets
(http://trac.wxwidgets.org/ticket/17737). This patch is a rollup of
the two merged commits which resolved the issue (the second patch
reverted the first attempt).

In the future, we should be able to remove this upon upgrading to the
next release of wxWidgets which includes this fix (if/when accepted).

Note that the build caches in Travis/Appveyor must be manually cleared
before this change is merged, as otherwise the cached, compiled
version of wxWidgets from prior builds will be used.

Fixes #27