Fix crash when opening a puzzle with debug plugin disabled.
Normally the luapuz library would be loaded by lua_openxword prior to
any use of Lua. However, on Mac (at least), the library would fail to
load because it would be loaded prior to the package path being set.
This was fine as long as the debug plugin was enabled, since the debug
plugin also loads luapuz and is initialized after the package path is
set. However, with the plugin disabled, nothing loads luapuz, and we
fail to push a Puzzle onto the Lua stack.

Swap the loading order to ensure that luapuz loads successfully, even
if the debug plugin is disabled. Also log a warning when the loading
fails to make future issues clearer. Note that we can't use
LogLuaMessage here without further tweaks because plugins are loaded
on a background thread, and setting the message count can't be done on
a background thread.

Fixes #60