Make XWord hidpi-aware.
The primary functional change here is that in the premake config, we
declare that the app is DPI aware, which turns off compatibility
rendering and allows the app to draw pixel-perfect UI on high-density
displays.

However, in any place that we currently specify a size in pixel units,
we must first scale the size to density-independent pixels to ensure
that UI elements are the same actual size regardless of the density.
While wxWidgets occasionally handles this automatically, in some
cases we must do the scaling ourselves. This commit handles most visible
places in the UI.

Specific notes:

- wxFormBuilder doesn't support hidpi yet. For Lua script UIs (generated
w/ py2lua), wrap any use of wxSize in a FromDIP call to scale it. For
C++ UIs, the only noticeable problem is with wxSpinCtrl; introduce a
wrapper class ScaledSpinCtrl which scales the provided size and use it
instead.
- Toolbar icons would need new image assets to scale properly. For now,
increase the size of the icon button while leaving the image at its
actual Pixel size, centering the image in the button.
- Calling font.SetNativeFontInfoUserDesc with an invalid string causes
an assertion error in the debug UI. Since this was only for backwards
compatibility with older XWord versions, we can probably safely delete
this.

Fixes #82