Fix metadata corruption when saving JPZ files.
When Puzzle.GetMeta was called for a key which didn't exist, it would
write an entry with an empty key and value into the metadata table. JPZ
serialization would convert this into an "<:anonymous />" XML tag which
is invalid in the JPZ specification and which breaks the Crossword Nexus
parser. By skipping this modification and leaving the metadata table
untouched, we prevent this invalid tag from being written.

See discussion in #66