Add YAML parsing support for Lua scripts.
Includes LibYAML 0.2.2 from https://pyyaml.org/wiki/LibYAML and lyaml
6.2.4 from https://github.com/gvvaughan/lyaml/releases.

LibYAML is unmodified except for yaml_get_version_string
in yaml/src/api.c. premake4 does not consistently escape quotes in
preprocessor macro definitions, so we can't properly set
YAML_VERSION_STRING to "0.2.2" as the standard Makefile normally would.
Instead we set it to 0.2.2 in the premake config and surround the value
in quotes in the source code.

lyaml contains both a C library and Lua scripts. The C library is built
in lua/lyaml, and the Lua scripts are added to scripts/libs/lyaml. The
library is unmodified except for lua/lyaml/ext/yaml/lyaml.h, where we
add a macro defining snprintf as sprintf_s on Windows (like yajl already
does) since snprintf is not available in older VS versions. This also
cannot be defined in the premake config as the definition must come
after importing stdio.h on newer VS versions to compile.

We also need to rename the C component of lyaml "c-yaml" because
otherwise the library is registered as "yaml" which tries to load the
library from the LibYAML yaml.dll.