Improve script and default config handling for distribution.
For scripts, migrate gen_nsis.lua to dist/gen_nsis.py since Python can
be run in stock AppVeyor environments whereas Lua is more challenging.
Edit the script to filter out plugin directories containing "_" or
"debug" which are not meant for release. Use the output of this script
directly when building the installer instead of copy-pasting it. Update
the zip command for Mac distributions to perform the same filtering on
the set of included scripts.

For the default config, move it to an explicit file instead of
generating it inline in the .nsi script, and bundle it on Mac as well.
Update the code to read from the default config if no config is present.

See #40