Initial work on Acrostic support.
This is a new extension to JPZ files being used by Outside the Box
puzzles and supported by the Crossword Nexus solver. It mostly follows
the JPZ standard, but uses the <acrostic> root XML tag instead of
<crossword> and requires that entries be propagated to their partner
squares with the same number.

This commit adds support for the new "Acrostic" grid type, enabling
loads/saves from the JPZ format. "clue" type squares are now supported -
these are used for the letters of each clue as well as punctuation in
the quote. Changes made to square contents (letters/flags) are
propagated to partner squares. The top-right text is changed to align
precisely with the clue number, as it was a pixel off before.

The primary remaining pieces for parity with Crossword Nexus are:

1.  Fixed navigation. The unconventional grid structure leads to strange
    arrow key and mouse click behavior when navigating through the grid.
    Navigation should consistently stay in the (only) Across direction
    and just move to the nearest square in that direction, switching
    words to the (only) word for that square if the word changes.

2.  Optionally, showing the completion message upon finishing the puzzle
    correctly, since for Acrostics this will generally include the fully
    formatted quote and author.