Add support for coded crosswords (aka "codewords").
Coded crosswords have no clues; instead, each square is numbered from
1-26, and each number represents a particular letter to be determined.
All squares with the same number contain the same letter.

Supporting these puzzles is quite similar to Acrostics in that we can
consider all the squares with the same number to be "partners" and
propagate changes to one square to all partner squares. We extend the
partner concept to support multiple partner squares instead of just one.

This provides a functional interface, but there are some drawbacks:

-  The cluelists still appear (Across/Down) despite being empty.
   Crossword Solver hides the unnecessary cluelist for this puzzle type.

-  The rendering performance is quite slow, probably because there are
   so many partner squares to update each time the cursor moves.

-  Crossword Solver shows a helpful letter bank indicating which letters
   have been used and which ones are remaining; this is missing from
   XWord.