Refine selection of initial square.
Previously, we focused on the first white square and an arbitrary word
in the Across direction containing that square.

Now, if there is either exactly one clue list, or a clue list labeled
"Across", we focus on the first square of the first word in that list.
Otherwise, we fall back to the same logic as before.