Respect clue list ordering from file and expand initial clue selection.
Change Clues from a map, which is ordered alphabetically, to a vector,
which retains insertion order. This should make no difference for
regular crosswords (and there is special logic to handle Across/Down
even if out of order), but retains the order from variety puzzle formats
which is more likely to be the intended viewing order.

Also expand the logic used to select the initial clue; previously, we
only selected the first clue's first square if there was one clue list.
We now do this even if there is more than one clue list. This is more
likely to be reasonable now that clue list ordering is maintained from
the original file.

Fixes #134