Improve handling of different capitalizations of "Across" and "Down".
When setting a ClueList, normalize them to the canonical "Across" and
"Down" as is used elsewhere in XWord. This fixes features like reuse of
existing settings for the Across and Down clue panes and saving the file
as an Across Lite .puz file.

Note that this results in the displayed clue list using the canonical
name and not the provided name, since MyFrame::ShowClues is built around
using the keys and not the titles. However, this is only a minor
cosmetic difference that seems harder to fix.