Propagate puzzle notes to PUZ files in the notepad.
Previously, extra metadata outside the standard PUZ notes field was
stored in a proprietary META section that is unknown to other
applications. This would result in an inconsistent experience, where a
JPZ file with visible notes from a non-standard field would be saved as
a PUZ that would then lose visible notes when viewed in Across Lite.

With this change, we remove support for the extra META section. Instead,
when saving to PUZ, we concatenate all metadata fields which are
considered to be notes (using section headers and separators) and insert
them as plaintext into the PUZ notes field. Of course this is not a
perfect rendering, but in the most common case of having one field for
notes but which might not be the <instructions> field as XWord expects,
this does a reasonable job.

There are two main functional changes (to my knowledge):

- Metadata fields not considered notes-like are now dropped entirely
  when saving as a PUZ, instead of being stored in the custom META
  section. This feels reasonable - when saving a JPZ as a PUZ, the most
  likely reason would be to open the PUZ in another app which won't
  support the META field anyway; if the intent is to save to a file that
  can be viewed in XWord itself, there's little reason to change format.
  But if desired, we could retain the META section as long as we omit
  any notes-like metadata when doing so (to avoid repeatedly
  concatenating it to notes each time we load/save the PUZ).

- Previously, when rendering JPZ notes, the normal "notes" section would
  have a slightly different shade of gray (#eeeeee) than the other
  sections (#dddddd). The difference is subtle and I can't discern a
  major reason for it; removing this allows us to simplify the logic in
  MyFrame::ShowNotes now that we've needed to push IsNotes() down to the
  Puzzle level.

See #140