Prevent focusing on black squares within a word.
We can only focus on squares which are white (unless the puzzle is
diagramless). So when a user presses an arrow key, even if the next
square in that direction is part of the current word, we should not
attempt to focus on it unless one of those conditions match.

Fixes #146