Fix shift-arrow navigation with annotation squares.
The intent is to navigate to the next square in a different word that
*could* be part of a word, even if it isn't. This is intended to be a
superset of all possible words, but can actually be stricter - for
example, if there is a three letter word with an annotation square in
the middle, neither of the two outer squares are considered potential
words because a word must have at least two letters.

So, in addition to accepting potential word candidates, we also accept
actual words, since some file formats support specifying words
explicitly rather than calculating them via this algorithm for potential
words.

See #146