Hack: check metatable in luapuz_push_tracked_object
I've seen errors for years like this:

    method GetSetSolution does not exist

from code like this, that works 90% of the time:

    square:SetSolution('.')

What seems to happen is that the square ends up with the wrong metatable
(often puz.Clue), even though the userdata pointer actually points to
memory that represents a puz::Square.

Best guess is there's some race condition around garbage collection and
removing objects from the luapuz tracked objects table. The pointer is
always correct, it's just an issue with the metatable. Checking for a
correct metatable when pushing an already tracked object does seem to
fix the issue for me, but it's definitely a band-aid, not a real fix.