Make shift-arrow logic more consistent.
Previously, when the arrow and focused direction were different, we'd
just move to the next white square in the arrow direction, even if there
was no word in the focused direction at that square. Instead, we should
keep searching in the arrow direction until we find a word (or possible
word) that could exist in the focused direction. The only difference is
that we avoid wrapping around the grid if the arrow direction and
focused direction aren't the same.

See #153