Initial support for reading and writing UTF-8 .puz files.
When reading a .puz, if the version code is 2.x (instead of 1.x), we
decode strings as UTF-8 instead of Windows-1252. This includes metadata
and clues, but not the grid itself.

When saving a .puz, if any metadata or clues uses characters which
cannot be encoded in Windows-1252, we save as a 2.x file and encode the
strings as UTF-8. This keeps backwards compatibility as wide as
possible, since not all .puz readers can handle 2.x files.

Loading/saving has been tested successfully with a basic UTF-8
character as used in the WaPo puzzle on 2021-06-20; the re-saved file
can still be loaded in Across Lite and has correctly-encoded text.
However, while the more complex example .puz with emoji clues can be
loaded, saved, and successfully loaded in Across Lite, the clues are
corrupted in the process. This needs more investigation, and may be a
bug in decode_utf8/encode_utf8 rather than the new logic here.