Fix regression in shift-arrow behavior on filled words.
In the old logic, we only moved to the start of a word explicitly when
the arrow direction and focused direction were different:

https://github.com/mrichards42/xword/blob/cd5d08cf9cd053025a6baa92ad13b818b1e3f7c8/src/XGridCtrl.cpp#L1674

The new logic applied this unconditionally. In most cases, this seems to
have the same behavior, in that if the word being moved to has any blank
squares, we move to the first blank in the word anyway. But if the word
being moved to is full, we should stay in line with the current focused
square to avoid jumping around the grid. This makes it easier to
navigate through a filled section of the grid.