Ensure XML characters are escaped when reading PUZ/RGZ.
XWord assumes that metadata and clue strings are valid XML/HTML when
rendering. This works fine for JPZ since the input is itself a valid XML
file, but may not be the case for other formats.

The clues and notepad were already being escaped when reading PUZ files
(and unescaped when writing them); we should do the same for the title,
author, and notes. Similarly, when reading RGZ files, we should escape
XML for any free-form text.