Support cell background colors in iPuz files.
iPuz supports two types of color declarations. Hex colors are
straightforward and fully supported. Puzzles can also use arbitrary
integers, which are app-defined values. For example, and app can define
0 = Black, 1 = Blue, 2 = Red, etc. This method seems less likely to be
used in practice, and it is non-trivial to define a palette that would
work well for arbitrary puzzles and user preferences. So for simplicity,
we treat 0 as Black per the spec recommendation, and everything else as
equivalent to "highlighted" (shaded gray).