More flexible solution character handling.
- Permit "*" as a valid grid character. While Across Lite doesn't
support it, other formats do. It can't be entered directly (since it is
bound as the shortcut for circling/uncircling a letter to match Across
Lite), but it can be entered as a rebus.

- When setting square text or solutions, never set the empty string,
even if ToGrid returns one (due to an invalid character, namely a black
square (".") or the symbol indicator ("[")). This ensures that bounds
checks pass in HasSolutionRebus.

See #185