Add Puzzle::LoadIpuzString
This is just for import scripts. Rather than having import scripts work
with Puzzle, Grid, Clue, and Square objects, they can just generate an
ipuz as json and load it that way.

Using ipuz as an internal interchange format means that (a) we should
make sure we support ipuz as fully as possible, and (b) we don't always
need to provide lua bindings for puz internals, since the ipuz loader
will handle that (e.g. we don't currently expose Square::m_bars in lua)