Fix thread safety issue in lua-yajl
lua-yajl used to keep a `js_null` _static_ pointer, which was set in
luaopen_luayajl (i.e. every time luayajl is required). When running from
multiple threads, luayajl is required and thus "opened" once for each
thread, causing `js_null` to be overwritten. Once this happens,
`js_null` is no longer the same pointer as `yajl.null`, since
`yajl.null` is owned by lua and is not shared between threads.

This shows up as `yajl.to_string` not being able to encode nulls, and in
fact, omitting them entirely, generally resulting in invalid JSON, such
as encoding `{"key": null}` as just `{"key"}`.

The fix is to get rid of the shared `js_null`, and instead compare with
the `yajl.null` value stored in the lua registry.