Add support for coded ipuz files.
This is an unofficial extension documented at
https://crosswordnexus.com/ipuz/coded/, but it's as simple as a custom
type on an otherwise normal crossword. If/when ipuz officially supports
the format, we would just need to add the official kind.

Generalize kind parsing by looping in reverse from most to least
specific until we find a type that we know how to handle.